==> benchmarks/lupus/scripts/03b-prepare-sim-replicates.R <==
## Generates simulated DE data from a set of mock data. The input data is
## expected to be a nested list of SingleCellExperiment objects for a *single
## celltype* representing different replicates of mock group assignment with
## various subsamples nested within depending on the number of cells per
## patient. This data is generated by scripts/03-prepare-mock-replicates.R


# Command line arguments --------------------------------------------------

library(argparse)
parser <- ArgumentParser()
parser$add_argument("-v", "--verbose",
    action = "store_true",
    help = "Print verbose output. Optional flag."
)
parser$add_argument("-i", "--infile",
    type = "character",
    default = "data/lupus-SCE_list-ncM-mock_replicates.rds",
    help = "Input data file name. Default: \'%(default)s\'."
)
parser$add_argument("-o", "--outfile",
    type = "character",
    default = "data/lupus-SCE_list-ncM-sim_replicates-prop_DE-0_01.rds",
    help = "Output results file name. Default: \'%(default)s\'."
)
parser$add_argument("--prop_DE",
    type = "double", default = 0.05,
    help = "Fraction. Proportion of genes to simulate as DE."
)
parser$add_argument("--countsimQC",
    action = "store_true",
    help = "Generate countsimQC report. Optional flag."
)
args <- parser$parse_args()
verbose <- args$verbose

if (verbose) {
    message("Using input file: ", args$infile)
    message("Using output file: ", args$outfile)
    message("Using prop_DE: ", args$prop_DE)
}

## Directory setup
here_root <- "benchmarks/lupus"
here::i_am(file.path(here_root, "scripts", "03b-prepare-sim-replicates.R"))

in_file <- here::here(here_root, args$infile)
out_file <- here::here(here_root, args$outfile)

if (!file.exists(in_file)) stop("Could not find input file ", in_file)
stopifnot(dir.exists(dirname(out_file)))

prop_DE <- args$prop_DE
stopifnot(length(prop_DE) == 1)
stopifnot("prop_DE is not between 0 and 1" = dplyr::between(prop_DE, 0, 1))

## Libraries
suppressPackageStartupMessages({
    library(SingleCellExperiment)
    library(DDCompanion)
    library(countsimQC)
    library(purrr)
})

seed <- 20220505


# Load data ---------------------------------------------------------------

## Load data
if (verbose) message("Loading data...")
sce_list <- readRDS(in_file)

## Check for expected data structure
stopifnot(all(map_lgl(sce_list, is, "SingleCellExperiment")))


# Simulate DE -------------------------------------------------------------

if (verbose) message("Simulating DE...")

## Simulate 'prop_DE' of DE genes for each mock replicate
set.seed(seed)
out <- map(sce_list, function(x) {
    simulate_DE(x, groups = x$mock_group, prop_DE = prop_DE)
})

## Compute t-SNE for reach simulated replicate
out <- lapply(out, compute_tSNE)

## Generate countsimQC reports
if (args$countsimQC) {
    ## Get celltype from infile name
    ct <- sub("^data/lupus-SCE_list-(.+)-mock_replicates\\.rds$", "\\1", args$infile)
    stopifnot(length(ct) == 1)

    countsimQC_dir <- here::here(here_root, "data", paste0("countsimQC-", ct))
    fs::dir_create(countsimQC_dir)
    p <- sub("\\.", "_", prop_DE)

    iwalk(out, function(x, i) {
        sim_counts <- as.matrix(counts(x))
        ref_counts <- as.matrix(counts(sce_list[[i]]))
        out_file <- sprintf("lupus-countsimQC-%s-prop_DE-%s.html", i, p)

        countsimQCReport(
            ddsList = list(
                "simulation" = sim_counts,
                "reference" = ref_counts
            ),
            outputFile = out_file,
            outputDir = countsimQC_dir,
            description = "Gene-swapping simulation for Lupus data",
            calculateStatistics = FALSE,
            forceOverwrite = TRUE,
            quiet = verbose
        )
    })
}


# Export output -----------------------------------------------------------

if (verbose) message("Exporting output data...")
saveRDS(out, file = out_file)


if (verbose) message("Done.\n")
