==> benchmarks/lupus/scripts/04b-run-DE-analyses-pb.R <==
## Runs mock DE analyses on the provided input data set using the specified DE
## methods. The input data is expected to be a nested list of
## SingleCellExperiment objects for a *single celltype* representing different
## replicates of mock group assignment with various subsamples nested within
## depending on the number of cells per patient. This data is generated by
## scripts/03-prepare-mock-replicates.R


# Command line arguments --------------------------------------------------

library(argparse)
parser <- ArgumentParser()
parser$add_argument("-v", "--verbose",
    action = "store_true",
    help = "Print verbose output. Optional flag."
)
parser$add_argument("-i", "--infile",
    type = "character",
    default = "data/lupus-SCE_list-ncM-mock_replicates.rds",
    help = "Input data file name. Default: \'%(default)s\'."
)
parser$add_argument("-o", "--outfile",
    type = "character",
    default = "results/lupus-ncM-mock_results-muscat.rds",
    help = "Output results file name. Default: \'%(default)s\'."
)
parser$add_argument("--method",
    type = "character", default = "muscat",
    help = "DE methods to be run on the input data. Default: \'%(default)s\'."
)

args <- parser$parse_args()
verbose <- args$verbose

if (verbose) {
    message("Using input file: ", args$infile)
    message("Using output file: ", args$outfile)
    message("Using DE method: ", args$method)
}

## Directory setup
here_root <- "benchmarks/lupus"
here::i_am(file.path(here_root, "scripts", "04b-run-DE-analyses-pb.R"))

in_file <- here::here(here_root, args$infile)
out_file <- here::here(here_root, args$outfile)

if (!file.exists(in_file)) stop("Could not find input file ", in_file)
stopifnot(dir.exists(dirname(out_file)))


## Libraries
suppressPackageStartupMessages({
    library(SingleCellExperiment)
    library(purrr)
    library(DDCompanion)
    library(scuttle)
    library(limma)
})

print(packageVersion("DDCompanion"))

# Prepare data ------------------------------------------------------------

## Load data
if (verbose) message("Loading data...")
sce_list <- readRDS(in_file)

## Check for expected data structure
stopifnot(all(unlist(map(sce_list, is, "SingleCellExperiment"))))

sce_list <- map(sce_list, prepare_SCE,
    group_id = "mock_group",
    subject_id = "ind_cov",
    cluster_id = "ct_cov"
)

## Binarize the data
## TODO: If we wish to also run an edgeR DGE analysis on the counts, we should
## also retain the original counts
sce_list_bin <- lapply(sce_list, function(element){
    assay(element)[assay(element)>=1] <- 1
    element$data_type <- "sce"
    return(element)
})

## Pseudobulk the data
pb_list_bin <- lapply(sce_list_bin, function(element){
    pb_bin <- aggregateAcrossCells(element,
                                   ids = element$ind_cov)
    colnames(pb_bin) <- paste0("identifier_", 1:ncol(pb_bin))
    pb_bin$data_type <- "pb"
    return(pb_bin)
})

# Run DE analyses ---------------------------------------------------------

## Arguments for DDCompanion::run_de_method()
tmp <- sce_list_bin[[1]]  # same for all replicates
if (nlevels(tmp$batch_cov) == 1) {
    formula <- ~ group_id
} else {
    formula <- ~ batch_cov + group_id
}

## Loop over nested list and run DE method on each SCE
if (verbose) message("Runnig DE analyses...")
out <- map(pb_list_bin, run_de_method,
    method = args$method,
    formula = formula,
    coef = "group_idB",
    BPPARAM = BiocParallel::MulticoreParam(workers=4),
    combined = FALSE
)

## Export results
if (verbose) message("Exporting results...")
saveRDS(out, file = out_file)

if (verbose) message("Done.\n")
